Fixing some issue with validatio and testing. Creating some new clases.

+++ Sergio_Bernal_Code/HackBank/src/main/java/com/hackbank/persistence/models/AccountType.java
@@ -2,22 +2,22 @@ package com.hackbank.persistence.models;
 
 public class AccountType {
 	
-	private int id;
+	private byte id;
 	private String name;
 	
 	public AccountType() {}
 
-	public AccountType(int id, String name) {
+	public AccountType(byte id, String name) {
 		super();
 		this.id = id;
 		this.name = name;
 	}
 
-	public int getId() {
+	public byte getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(byte id) {
 		this.id = id;
 	}
